initial reroute functional, renamed app files
