+++ app/api/auth_routes.py
@@ -41,7 +41,7 @@ def login():
         # Add the user to the session, we are logged in!
         user = User.query.filter(User.email == form.data['email']).first()
         login_user(user)
-        return user.to_dict()
+        return user.to_dict(workspaces=True)
     return {'errors': validation_errors_to_error_messages(form.errors)}, 401
