+++ react-app/src/components/auth/LogoutButton.js
@@ -1,11 +1,14 @@
 import React from 'react';
 import { useDispatch } from 'react-redux';
+import { Redirect, useHistory } from 'react-router-dom';
 import { logout } from '../../store/session';
 
 const LogoutButton = () => {
   const dispatch = useDispatch()
+  const history = useHistory();
   const onLogout = async (e) => {
     await dispatch(logout());
+    history.push('/')
   };
 
   return <button onClick={onLogout}>Logout</button>;
